+++ changelog.md
@@ -14,3 +14,21 @@ added - movement() function
 added - options to walk's sub menu "N,S,E,W" in movement()
 fixed - movement so player can not walk off "map"
 
+v2.0 
+GitHub Name: Map---part-2 
+created - empty list to function as an inventory of objects 
+created - objects without a dictionary 
+added - conditional statements to deal with objects
+
+v3.0 
+GitHub 
+Name: Map---Dictionarys 
+Name: RPG: Castal Adventure! 
+added - dictionary for tile/map attributes (description strings) 
+fixed - room description for main menu 
+updated - object data to a dictionary 
+added - 'look' as a main menu option 
+added - inspectRoom() function to deal with objects in the game 
+added - a fuction to deal with the chest object 
+added - a function to deal with the key object 
+added - docStrings for each of the functions